% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplpy.R
\name{ampl_env}
\alias{ampl_env}
\title{AMPL Environment}
\usage{
ampl_env(ampl_path = "/Applications/ampl")
}
\arguments{
\item{ampl_path}{Character Path to AMPL executable}
}
\description{
Creates a new AMPL model
}
\details{
As this function instantiates a new background AMPL
process, if you only require one model environment at
a time, use \code{reset} to clear the current process
instead.
}
