% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{display}
\alias{display}
\title{Display}
\usage{
display(ampl, entity)
}
\arguments{
\item{ampl}{\code{amplpy.ampl.AMPL} object}

\item{statements}{Character. The display statements to be fetched.
Can return more than one model entity provided they are defined
over the same indexing sets (See Details.)}
}
\value{
data.table containing the resulting display command
in tabular form.
}
\description{
Low-level function equivalent to the AMPL call
}
\details{
\code{display ds1, ..., dsn;}

where \code{ds1, ..., dsn} are given as the parameter \code{statements} here.

As AMPL only returns one table, the operation will fail if the
results of the display statements cannot be indexed over the same set.
As a result, any attempt to get data from more than one set, or to get
data for multiple parameters with a different number of indexing sets
will fail.
}
