% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampl.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate}
\usage{
evaluate(ampl, amplstatements)
}
\arguments{
\item{ampl}{\code{amplpy.ampl.AMPL} object}

\item{amplstatements}{A collection of AMPL statements and declarations to
be passed to the interpreter.}
}
\description{
Parses AMPL code and evaluates it as a possibly empty sequence of AMPL
declarations and statements.
}
\details{
As a side effect, it invalidates all entities (as the passed statements
can contain any arbitrary command); the lists of entities will be
re-populated lazily (at first access).

The output of interpreting the statements is passed to the current
OutputHandler (see getOutputHandler and setOutputHandler).

By default, errors are reported as exceptions and warnings are printed
on stdout. This behavior can be changed reassigning an ErrorHandler
using setErrorHandler.
}
