% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampl.R
\name{set_option}
\alias{set_option}
\title{Set Option}
\usage{
set_option(ampl, name, value)
}
\arguments{
\item{ampl}{\code{amplpy.ampl.AMPL} object}

\item{name}{Character Name of the option to be set (alphanumeric without spaces)}

\item{value}{Character The value the option must be set to.}
}
\description{
Set an AMPL option to a specified value.
}
