% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampl.R
\name{set_solver}
\alias{set_solver}
\title{Set Solver}
\usage{
set_solver(ampl, solver_path)
}
\arguments{
\item{ampl}{\code{amplpy.ampl.AMPL} object}

\item{solver_path}{Character Path to the solver executable}
}
\description{
Set the current model solver.
}
