% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampl.R
\name{use_objective}
\alias{use_objective}
\title{Use Objective}
\usage{
use_objective(ampl, objective)
}
\arguments{
\item{ampl}{\code{amplpy.ampl.AMPL} object}

\item{objective}{Character name of the objective function to be solved}
}
\description{
Sets an objective function as the current model context
}
\details{
Call this function if your model has multiple objectives defined in
order to switch between them. This only alters the state of the model
environment and doesn't execute the solver.
}
